fix: update to latest API's
Major breaking change to the pattern of how pathtracer works so this was updated.

Most of the API still works but some things moved.

Depreciated using the word renderer as it confuses with the default renderer, instead use `pathtracer`

Some options removed as now handled by three and DREI

Three updated to 168 R3F to 8.17.8
All other libraries updated to latest

Major version bumped to 1.1.0